fix: Fix GraphQL field name mismatches causing CI/CD failures
- Update query/mutation field names to match GraphQL schema (camelCase)
- Remove incorrect camelCase conversion from GraphQL template
- Fix Float vs Int type mismatches for limit/offset parameters
- Preserve snake_case for arguments and field names as required by schema

The GraphQL schema expects:
- camelCase for operation names (getApiKeys, createDiagram, etc)
- snake_case for arguments (api_key_id, diagram_id, etc)

+++ dipeo/models/src/frontend/query-definitions/executions.ts
@@ -75,8 +75,8 @@ export const executionQueries: EntityQueryDefinitions = {
       type: QueryOperationType.QUERY,
       variables: [
         { name: 'filter', type: 'ExecutionFilterInput' },
-        { name: 'limit', type: 'Int' },
-        { name: 'offset', type: 'Int' }
+        { name: 'limit', type: 'Float' },
+        { name: 'offset', type: 'Float' }
       ],
       fields: [
         {
